+++ Changelog.md
@@ -28,5 +28,6 @@
 - [Add Warden and Dragon model to Boss Crusher](https://github.com/TheStaticVoid/StaTech-Industry/issues/422) - There's more planned for this in the future
 - [Added a tooltip to all custom MI machines](https://github.com/TheStaticVoid/StaTech-Industry/issues/426)
 - [Added packer recipe for TechReborn Scrap Box](https://github.com/TheStaticVoid/StaTech-Industry/issues/430)
+- [Added assembler and packer recipes for AE Fluix Cables](https://github.com/TheStaticVoid/StaTech-Industry/issues/431)
 - [Fix modernfix config](https://github.com/TheStaticVoid/StaTech-Industry/issues/420)
 - [Fix a recipe issue with ME Cables from MI datagen](https://github.com/TheStaticVoid/StaTech-Industry/commit/d18d1b6a0f2ad2f01b5763c1f61ea1d8e120c554)